pyproject.toml
[project]
name = "coderipple"
version = "0.1.0"
description = "Git历史分析工具 - 追溯bug是否影响目标版本"
authors = [
    {name = "Lin Li", email = "ara.lilin.fortune@google.com"}
]
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "gitpython>=3.1.0",
    "click>=8.0.0",
    "pyyaml>=6.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0"
]

[project.scripts]
coderipple = "src.cli.main:cli"

[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
